Fixed missing `apple.security` to SecurityTest. Issue: #67
Also fixes a problem related to running JUnit Tests from IntelliJ on Mac
systems. IntelliJ resides in a folder `IntelliJ App` (mind the space)  which executes the
JUnit tests and adds therefore its own Jars to the classpath. When
converted from an URL into an URI we recieve an illegal character
exception. The fix was to UrlEscape the path variable within the `newUrl` method.

Signed-off-by: Dennis Rippinger <dennis.rippinger@gmail.com>