use Johnson algorithm for cycle detection
The existing implementation of cycle detection has run time growing at least exponentially in the number of nodes and vertices.
This implementation of the algorithm is known to have complexity O((n+e)*(c+1)) where n, v and c are the numbers of vertices, edges, and elementary cycles respectively.

We map our complex domain objects to a primitive representation (`int` for nodes and `int[]` for edges) to avoid pitfalls with complex data types like hashcode overhead. After we have executed Johnson's algorithm on this optimized primitive structure we do a re-mapping to our domain objects and process the result.

As a basic protection from regression we add an integration test which does not terminate within 5 minutes for the previous implementation and now terminates within seconds.

Thanks @codecholeric for large improvements on the code, discussions,
and review!

Signed-off-by: Philipp Vollmer <briefe@kebes.de>