Clarify difference between nested and inner classes
JLS: "A nested class is any class whose declaration occurs within the
body of another class or interface." [1]

JLS: "An inner class is a nested class that is not explicitly or
implicitly declared static." [1]

Before this commit, the method JavaClass#isInnerClass returned true for
all nested classes.

This commit changes JavaClass#isInnerClass that it only returns true for
inner classes.

This commit also introduces JavaClass#isNestedClass that returns true
for all nested classes.

[1] https://docs.oracle.com/javase/specs/jls/se8/html/jls-8.html

Signed-off-by: Roland Weisleder <roland.weisleder@googlemail.com>