fix accidentally invariant predicate parameter types
By accident some DescribedPredicate<T> method arguments of public API methods had been defined in an invariant way, e.g. DescribedPredicate<JavaClass> instead of DescribedPredicate<? super JavaClass>. Method parameters defined this way prevent an easy reuse of predicates for super types, even though a DescribedPredicate for a supertype always also is a DescribedPredicate for a subtype. To take the above example, we want users to be able to pass a DescribedPredicate<HasName> to any method that takes a DescribedPredicate<JavaClass>, and this is only possible if the parameter type is defined as DescribedPredicate<? super JavaClass>.
I have added an ArchTest to PublicAPIRules to ensure this. Unfortunately ArchUnit does not support type parameters yet, thus the Reflection API has to be used for the type parameter analysis.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>