exclude all self references from dependencies
So far we have only filtered out self dependencies from accesses, it actually can happen in other cases as well though (field or parameter of own type, etc.). Thus we should move this logic closer to the creation of dependencies since it is an invariant of `Dependency`. The signature returning `Optional<Dependency>` is not terribly pretty and expressive, but since it is only internal code, I considered it okay (`Dependency` is not a terribly complex class, so I think it is quick to find the cases when the result is `Optional.absent()`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>