upgrade Spotbugs (plugin-4.0.5/tool-4.0.1)
Somehow Spotbugs 3.x seems to have broken dependencies meanwhile, so I have upgraded it to the current version. Seems like the plugin API has changed meanwhile and the source sets cannot be manipulated anymore, thus I have adapted it to disable the test tasks (with all the many test examples we need for tests, Spotbugs reports a lot of useless warnings for tests, so we only enable it for production code). Also there now seems to be automatically one task per source set, so we do not need to manually add `sourceSets.api` anymore.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>