support type variables as upper bounds of other type variables
We now support signatures like `Foo<S, T extends S>`, where one type parameter is bound by another type parameter. What makes this difficult is the fact that `extends S` in this case is a reference to the original type parameter `S` (same holds for the Java Reflection API). I.e. in a case like `Foo<S extends String, T extends S>` it must be possible to retrieve `typeParameters[1].upperBounds[0].upperBounds[0]` to be `java.lang.String`. So to build the type parameter `T`, we need the type parameter `S` already built so we can reference it.
To support this, I have split the creation of type parameters into `build()`, which will skip the bounds (since we need the built type parameters for this), and `complete` where the bounds are finished and set into the type parameters (thus unfortunately removing immutability, but I do not think you can have this in such a chicken and egg scenario with Java capabilities; in any way the mutability is package private, from an API point of view this object will be immutable like all ArchUnit domain objects). Unfortunately this is a fairly recursive thing, so each builder in the type parameter bounds tree needs all constructed type parameters, because it might have bounds or assigned type parameters somewhere down the line that are type parameters themselves.
Within the import process we now need one more level of indirection, because we must decide if this is a new type to build (like a parameterized type) or if it simply references another type parameter and thus needs to return an already constructed one. Therefore there is now `JavaTypeCreationProcess` to encapsulate these two cases (technically it would have been possible to hack `TypeVariableBuilder` to return an existing type variable on build, but I have decided against it, since it would be a violation of the builder pattern and thus surprising).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>