prepare release `0.14.0`
Consistently set version to `0.14.0` in user guide by using an attribute reference to `{revnumber}`. Also upgraded dependencies of AsciiDoctor to current versions and included `/docs` into the core build, so attributes like the version are automatically consistent and we have a task in the root project to create the user guide.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>