fix broken classpath scanning for manifest classpaths
Tools often have an option to specify the classpath as manifest entry, i.e. to run tests the tool will create an empty JAR only containing a manifest with an `Class-Path` attribute containing the actual classpath. This is useful if the classpath is long and the platform (e.g. Windows) does not support a sufficiently long command line, thus limiting the length of the classpath supplied on the command line.
Unfortunately the recent fix for Android, switching from loading the classpath through the respective Classloader to resolving the classpath manually, did not take this case into account anymore and it slipped through the tests, since no environment would specify the classpath via manifest. Now we also resolve those manifest attributes when manually resolving the classpath.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>