only check if MD5 is enabled once per import
It does not make sense to continuously read `md5EnabledInClassSources` for every single class. Instead we now read the configuration once for all the classes to import and pass the configured value through with the source URI. This means that between separate runs of `ClassFileImporter.import...()` the property can be changed, but not while the import is running.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>