rename `JavaGenericTypeImporter` -> `JavaClassSignatureImporter`
We should separate importing the signature of fields or methods in the future, but it makes sense to also integrate generic superclass and interface handling here, because the scope is very similar. I have thus renamed the class to be a "class signature importer", which is less generic than "generic type importer", but leaves room to handle other class signature related parsing. For now it continues to only handle type parameters, but in the next step we can extend this to generic superclass parsing.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>