remove redundant search for call target
We have already checked that the call is not directly to `parent` before (because otherwise the respective member candidate would already have been an exact match). Thus it is futile to check if the call targets `parent` and we can leave it out from our search and remove `Sets.union(..)`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>