add generic superclass to `JavaClass`
Unfortunately signature handling is a little bit inconsistent. If the type is not generic the signature is `null` and we need to rely on the superclass name passed to `visit(..)`. If the type is generic we will now derive the raw superclass from the erasure of the generic superclass. This seems logically consistent. It could however break existing behavior, if for some reason our assumption is wrong and the erasure of the generic superclass is not always equivalent to the raw superclass whose type name is passed as `superclass` string to `visit(..)`.
To be consistent with `HasType`, which offers `getType()` and `getRawType()` we have decided to name the methods `JavaClass.getRawSuperclass()` for the raw superclass and `JavaClass.getSuperclass()` for the generic superclass (i.e. the generic case will be the "default" case as opposed to the reflection API where this is always named explicitly as "generic type").

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>