prevent creation of parameterized types without type arguments
To be consistent with the Reflection API we should not create a `JavaParameterizedType` with an empty list of type arguments for generic type signatures. So far for `List<? super String>` we would e.g. create a `JavaParameterizedType{String}` as upper wildcard bound, even though `String` obviously has no type parameters. The Reflection API would return the `Class<String>` directly as upper bound in this case, so we should return the `JavaClass{String}` directly here.
It feels a little fishy to have `JavaParameterizedTypeBuilder` return a `JavaClass`, if there are no type arguments (inconsistent pattern, if `FooBuilder` always returns `Foo` in all other cases), but due to the visitor pattern it is not trivial to know ahead of time which builder is the correct one to use. Since `JavaParameterizedTypeBuilder` is an implementation detail and usually encapsulated by referring to a generic `JavaTypeBuilder`, I still decided to go this simple way. It would be fair to simply give `JavaParameterizedTypeBuilder` a name that would reflect that both `JavaParameterizedType` and `JavaClass` could be created, but I could not come up with a good name.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>