run artifact checks on every build
Since the check takes only a couple of ms we can make it part of the regular build. This will enable us to detect wrongly repackaged dependencies early on and fix them inside of the respective PR instead of at release time. We should still run the check on the release artifact as well, just to be on the safe side.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>