fix `getName()` of `JavaWildcardType` to include bounds
This should behave like the Reflection API. While for a `WildcardType` in the Reflection API there is no `getName()` method, we want `getName()` to match `Type.getTypeName()` which is available from Java 8 on. However `getTypeName()` includes the bounds in the name, similar to how the `toString()` method of `JavaWildcardType` already works. I have thus adjusted the name to contain the bounds of the wildcard.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>