fix `getName()` of `JavaParameterizedType` to include type arguments
`getName()` of `JavaParameterizedType` should match the Reflection API's `ParameterizedType.getTypeName()`.
Since we want to compare these two methods in the test and the method does not exist in Java versions < 8 I have added the test to the `jdk9test` source set for now. Once we drop Java 7 support we can move it back to standard `test` source set.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>