rename generic type parameters of test classes
`SUPER` and `SUB` were fitting names for the superclass and interface tests, but for the field test they don't match very well. I chose to give the `GenericFieldType` a quite generic name `T` and name the type variable we're interested in `OF_CLASS` to emphasize that it is defined as type parameter of the class defining the field.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>