use data provider defaults in ClassFileImporterGenericInterfacesTest
For these one-to-one associations where we always have pairs of data providers and test methods we should use the default data provider naming convention of prefixing the test with `test_` and the data provider with `data_`. This automatically enforces a naming convention and eliminates one possible error where a string can run out of sync with a method name.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>