prepare ArchUnit `Optional` for migration to Java 8
Once we drop support for Java 7 we can also drop ArchUnit's own `Optional` type and use the standard `java.util.Optional` instead. To make that transition easier it makes sense to already adjust the ArchUnit `Optional` to the Java 8 `Optional` as much as possible (originally it was modelled close to the Guava `Optional`).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>