add generic parameter types to `JavaCodeUnit`
Note that raw and generic parameter types do not behave the same way in all cases. Specifically for constructors of inner classes that obtain a reference to their enclosing class. In this case the Reflection API returns the enclosing class as first raw parameter of the constructor, while the generic parameters (in case there are any) do not contain the enclosing class. Thus I decided to keep both raw and generic parameters in `JavaCodeUnit` separately, while `getParameters()` will return the raw parameters (and thus one extra parameter for inner classes) in non-generic cases. This does not feel great, but the alternatives seem even worse. At least this way it is consistent to the Reflection API. If we choose a different path there will most likely be even more confusion.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>