add method and constructor references to `JavaClass`
This will add a way to detect method references (e.g. `SomeClass::call`) and constructor references (e.g. `SomeClass::new`) within classes. Detecting this is quite challenging because it is hidden behind a `invokeDynamic` instruction in the bytecode. And `invokeDynamic` in turn can be used for many more things than method/constructor references. The way we can detect this is the following:

* analyse `invokeDynamic` instructions within the bytecode by hooking into ASM's `visitInvokeDynamicInsn`
* look for any instructions that target the JDK's `LambdaMetafactory` as a bootstrap method owner
* if the second argument of the bootstrap method on the metafactory is a method handle it is a candidate for a method reference (this seems to be always the way, by convention)
* at this point we have two possibilities: a) the candidate is a method/constructor reference or b) it is a synthetic lambda method (this is a static method the compiler generates whenever a lambda, e.g. `() -> foo + bar`, is declared)
* if the method is not a lambda method (we can detect this by a naming convention, e.g. the method is called sth. like `lambda$someName$0`) we found a method/constructor reference

We introduced a new domain object hierarchy branch, i.e.

```
                               /-- JavaMethodCall
                   /-- JavaCall
                  /            \-- JavaConstructorCall
JavaCodeUnitAccess
                  \                         /-- JavaMethodReference
                   \-- JavaCodeUnitReference
                                            \-- JavaConstructorReference
```

The reason is that each level of the hierarchy might be interesting for users and provide value. A user might be interested in any access to any code unit (e.g. check all parameters of any code unit that has been targeted, be it a method call or a constructor reference, etc.). They might also be interested in any real call of a method/constructor (after all, a reference doesn't mean this method is really called at that moment, in the end we only define a pointer to a method, not call it). Or they might be interested to see only references to declare some assertions which methods/constructors may be referenced. Thus, we decided it makes sense to expose each of these domain objects from the public API.

Note that by integrating references into the domain object hierarchy of `JavaAccess` we automatically have all references also registered as dependencies (i.e. getting picked up by `layeredArchitecture()` and similar).

Signed-off-by: Korbinian Singhammer <korbinian.singhammer@tngtech.com>
Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>