improve "release"/"docs" naming in tasks
`releaseArchUnit` was a misleading name, because it actually only took care of keeping the website and the examples up to date. While publishing of the artifacts would obviously also be expected as part of a release. We now call this `publishDocs`, which refers to all kinds of documentations like website+user-guide/examples. In turn I have renamed the old `updateArchUnitDocs` task to be named `updateArchUnitWebsite`, because that is more descriptive of what it actually does.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>