automate version calculation and release preparation
We can further simplify the release process by doing all version calculation automatically. Also doing the update related stuff in Gradle might be a good idea in case we need more powerful logic in the future (version calculation is already one such example, but I think if we already do that in Gradle, we might as well do the substitution of the relevant places in Gradle).
To calculate the version, we can pull the last release version out of the README and the new release version should always be the current SNAPSHOT version with the dropped suffix.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>