fix `URISyntaxException` if `ClassLoader` returns unencoded URL chars
It seems like some OSGI `ClassLoaders` would return resource URLs that don't have correctly encoded special characters (in particular spaces). This leads to `URISyntaxException` if `url.toURI()` is called on those `URLs`. Note that the default JDK `ClassLoaders` don't show this behavior, i.e. the issue is not reproducible with standard `ClassLoaders` because in these cases spaces are correctly encoded.
Since this is not the standard case I decided to add a `try` `catch` block to solve this issue. In the catch block we try to take apart the URL and pass the parts to the structured constructor of `URI`, which will then in turn encode special characters in the path correctly.

Issue: #683
Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>