introduce `JavaCodeUnit.Parameter` to encapsulate parameter details
Same as the Reflection API we can offer a more powerful `Parameter` object, which can tell its type as well as its annotations in the future. Having a nice consistent view on a single parameter to easily match type and annotations makes sense for writing ArchRules.
Note that there are some cases where we cannot safely determine the parameters. These are cases where the compiler adds some synthetic parameters. We can resolve it for common cases (like constructors of inner classes have the outer class as first raw parameter, even though it is missing from the generic signature). But in general there is no way to do this safely, because it can even differ from compiler to compiler, and I could not find any way to resolve the flag SYNTHETIC of a parameter with ASM.
 For example local classes will have parameters from the outer context appended to the constructor signature, so we can also not assume synthetic parameters are always in the beginning, nor a fixed number. Altogether, there is just no safe way to do it for all corner cases, so I decided to compensate the two known common cases (inner class constructor and enum constructors) and otherwise fall back to creating parameters just from the raw parameter types. This way they will at least be consistent and even the Reflection API has to do this, as it is also not possible to determine this in all cases with Reflection (compare `java.lang.reflect.Executable.getParameters()`).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>