remove indirection `ClassesByTypeName` from import
Originally this was supposed to apply the interface segregation to only pass on what is really needed for the following builders. However annotations already need more to build and thus need the full `ImportContext`. I feel that `ImportedClasses` is still okay as an interface, since the scope is restricted to the import process. On the other hand we can then also refer to `ImportedClasses` when building annotations in a following step.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>