add short sleep time after executing commands
I could not find out why this happens, but for example if we call `git checkout -b` immediately followed by `git rev-parse` from Gradle the output will be empty. I tried to reproduce it locally and with plain Git but couldn't. It only happens during the release job when it runs as GitHub action, so it is quite hard to debug. By my understanding the process should be finished if `waitFor()` returns (and the exit code is even 0), but somehow this does not work reliably. So for now I decided to go the quick and dirty way and simply add a short sleep time after executing shell commands.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>