replace cross-project config `allprojects` config by plugin
Meanwhile, there is quite some cross-project configuration between subprojects of ArchUnit. This is not only considered [bad practice](https://docs.gradle.org/6.9.2/userguide/sharing_build_logic_between_subprojects.html#sec:convention_plugins_vs_cross_configuration) but also causes real pain in maintainability, since dependencies between the projects are hard to understand and configuration order causes more and more problems. We will thus refactor the build logic to follow a plugin-based approach as described in the Gradle user manual. As a first step we can introduce a `base-` and a `java-conventions` plugin to be applied to projects, which replaces the `allprojects` block in the root build file.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>