upgrade Spotbugs
I'm not sure if the `toolVersion` before really worked, because simply setting it to `4.5.2` does not seem to have pulled in the new version. Only combining it with also setting the dependency of the `spotbugs` configuration seems to work. However, Spotbugs now reports a lot of "private member is exposed" violations. Since ArchUnit heavily uses immutable collections and objects, where exposure is not critical, this inspection is basically useless since it has way too many false positives. It would also not be very convenient to maintain in the future if we have an exact list of ignores, thus I have disabled those inspections globally.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>