fix `ExamplesIntegrationTest` on JDK target compatibility version 7
When the whole classpath is imported there are some more accesses from `org.jcp..` to `java.security..` for JDK 7. Since the point of this test is to find the violations in the app code I simply added this to the accepted origins (after all an additional package won't cause any pain for other JDK versions, if there is no call from there, since it's just the specification of possibly allowed packages).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>