replace `multiJdkTest` by Toolchains
Now that we can run the whole build with a specific JDK version we don't need the custom logic adding "multi JDK tests" anymore. The new approach is not only more thorough (it also compiles the classes under test with the respective JDK instead of only running the test VM with the respective version) but also simpler to understand.
Unfortunately we can't unify the config for JDK 7, since for JDK 7 we need to manually install the JDK and pass the location to the Gradle build, while for all other JDKs the Toolchain plugin can automatically provision the desired JDK(s). But since JDK 7 support will be dropped anyway in the near future this shouldn't cause too much pain.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>