fix JUnit 5 Maven Integration Test
The old version of the JUnit Vintage Engine could not process the new JUnit 4 version. This also shows, that unfortunately the Maven integration test has not been running on CI recently for newer Java versions, because otherwise it would have had to fail as well when we upgraded the JUnit 4 dependency. It looks like on CI the Maven integration test always ran with Java 7 for some reason (I don't know at what time exactly it broke, but we will fix it in a successive commit).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>