fix Gradle compile dependencies in module `:archunit`
The build has been a little shady, which showed e.g. because after a clean IntelliJ would sometimes crash on the first run after a `./gradlew clean`. The reason was that the dependencies between the test tasks and the JDK 9 compile were not set up correctly. Both `jdk9test` and `jdk15test` need the compiler output of `jdk9main` to work.
I've also hooked the `compileJdk9mainJava` task as dependency into the `assemble` lifecycle task, which is the correct umbrella task for the given nature. Altogether, after a `./gradlew clean` I could not reproduce the error in IntelliJ anymore, so hopefully the dependencies between compile/test are set up correctly now for the different source sets.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>