move resolving class hierarchy to resolution process
We should encapsulate all dependency resolution related logic into one dedicated place that only takes care of this one aspect. Note that we need to declare the test examples separately, because otherwise there is always compiler magic for inner/local classes that can screw up the test result (i.e. the test is green, but it's not because of class hierarchy resolution, but because it was resolved as some constructor argument type, etc.)

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>