implement containNumberOfElements for members
The method containNumberOfElements can be used to ensure that the `that`-clause matched a certain number of elements.
E.g. `classes().that().implement(SomeBusinessInterface.class).should().containNumberOfElements(lessThanOrEqualTo(1))`.
Until now, this method only existed for classes, but not for members.
With this commit we implemented it now also for members.

Signed-off-by: Hannes Oberprantacher <h.oberprantacher@gmail.com>