automatically resolve component types of array types
This will ensure that component types of array types (e.g. `String` for `String[]`) are fully imported. Otherwise, things like annotations or class hierarchy will be missing for types that are not directly imported and only referenced through component types.
In contrast to the other automatic resolutions I added this to the core resolve method in `ImportedClasses` because there are too many places where we also need to consider component types. Otherwise, this would never be robust for future changes (or at least always be a trip mine to overlook). To resolve the component type of any array type seems to be an intrinsic property of resolving any array type, so this seems fair to me.
I thought about adding a shortcut to skip resolving further down the component type hierarchy if we encounter a type that has already been imported, but this would only help for repeated multi-dim arrays of the same component type but different dimensions (otherwise the cached type would be used anyway, and we don't resolve again). So, to me the benefit seems quite unclear, and we should probably wait until some bottleneck has shown before prematurely optimizing this at the cost of readability.

Note that by this change the test `classes_not_fully_imported_have_flag_fullyImported_false_and_empty_dependencies` does not make sense anymore, because this can only happen if we resolve additional classes during the completion process. I do not think this will ever happen anymore (once we also resolve type parameters in the next step), because we strive to resolve all such dependencies ahead. So the cases appearing should only be fully imported or stub, which is already covered by the test `ClassFileImporterAutomaticResolutionTest.never_reports_stub_classes_as_fully_imported`. Likely to have the `FullCompletionProcess` within `JavaClass` will not make sense anymore, because we will not end up within the completion process needing a type that has not been resolved ahead. However, I am not 100% sure this is really the case under all conditions, so I will keep the `CompletionProcess` as it is.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>