never report stub classes as "fully imported"
So far the semantic of `isFullyImported()` was a little broken. When we create the stub class before the general completion process starts it will still run through all the completion steps and thus be reported as "fully imported". This is not correct though, a stub class we create can never be fully imported, because we are missing vital information from the bytecode. I have adjusted this to propagate the information that this class is a stub class and in that case never report it as "fully imported".

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>