disable enclosing and generic type resolution by default in tests
Unfortunately, this costs a lot of performance (the build time increased by 50%). This is not due to the resolution being terribly slow (in fact, when I compared it on big imports the performance even went up a little due to the new resolution mechanism), but because many tiny tests import `java.util.List`, etc., and each local class used in tests now leads to the test class being imported as enclosing class, including all of its dependencies. Thus, for these tests the number of imported classes simply increases quite a bit. Since this automatic resolution is irrelevant for most tests I have thus disabled it by default. Tests can still explicitly import the types of generic signatures or enclosing classes, or explicitly reconfigure the resolution behavior.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>