improve tests of `classes().should().containNumberOfElements(..)`
There are actually only two relevant cases to test: a) the predicate on the number evaluates to `true`, then the rule should pass and b) the predicate on the number evaluates to `false`, then the rule should fail. The former test was actually missing, the latter was grandly over-tested. Each of the predicates is tested in isolation (`equalTo`, `lessThat`, ...) so we can assume they work. Thus, the only thing we need to test is a generic predicate evaluating to `false`. We do not need to test the description of the rule as this is covered automatically by `Random...SyntaxTest`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>